+++ projects/angular-sdk/src/i18n/fr.ts
@@ -320,7 +320,7 @@ export const locale: I18n = {
             LOADING_HYPERWALLET: "Chargement d'Hyperwallet en cours...",
             LOADING_HYPERWALLET_ERROR: "Une erreur est survenue lors du chargement d'Hyperwallet",
             STEP: {
-                FIRST: "Type de remboursement",
+                FIRST: "Récapitulatif de la réservation",
                 SECOND: "Informations générales",
                 THIRD: "Information du paiement",
                 FOURTH: "Confirmation"
@@ -349,7 +349,13 @@ export const locale: I18n = {
             },
             CONFIRMATION_MESSAGE_1: "Votre vol a été remboursé and votre remboursement est en train d'être traité.",
             CONFIRMATION_MESSAGE_2: "Un mail de confirmation va être envoyé à l'adresse {{ userEmail }} avec les détails sur la manière d'utiliser votre crédit",
-            REFUND_DETAILS: "{{ amount }} en {{ method }}"
+            REFUND_DETAILS: "{{ amount }} en {{ method }}",
+            PASSENGER_TYPE: {
+                ADULT: "Adulte",
+                CHILD: "Enfant",
+                SEATED_INFANT: "Bébé assis",
+                LAP_INFANT: "Bébé sur les genoux"
+            }
         }
     }
 }